fix(mobile): Move hamburger menu to right of profile
The hamburger menu was positioned incorrectly next to the logo on mobile.
This moves it to the right side of the header, after the profile menu,
for a more intuitive mobile layout.
