+++ src/app/tips/[slug]/page.tsx
@@ -7,6 +7,7 @@ import { ExternalLink, Compass, RefreshCw, Command, MapPin, Cpu, MessageSquare, 
 import { AppLink } from '@/components/AppLink';
 import { MainNav } from '@/components/MainNav';
 import { UserMenu } from '@/components/UserMenu';
+import { MobileNav } from '@/components/MobileNav';
 import { TipBar } from '@/components/TipBar';
 import { PageContainer } from '@/components/PageContainer';
 import { useTimeRange } from '@/contexts/TimeRangeContext';
@@ -46,7 +47,10 @@ function GuideContent() {
           <PageContainer className="py-4">
             <div className="flex items-center justify-between gap-4">
               <MainNav days={days} />
-              <UserMenu />
+              <div className="flex items-center gap-3">
+                <UserMenu />
+                <MobileNav days={days} />
+              </div>
             </div>
           </PageContainer>
         </header>
