+++ src/app/users/[email]/page.tsx
@@ -9,6 +9,7 @@ import { StackedBarChart } from '@/components/StackedBarChart';
 import { TimeRangeSelector } from '@/components/TimeRangeSelector';
 import { MainNav } from '@/components/MainNav';
 import { UserMenu } from '@/components/UserMenu';
+import { MobileNav } from '@/components/MobileNav';
 import { UserProfileHeader } from '@/components/UserProfileHeader';
 import { AdoptionBadge } from '@/components/AdoptionBadge';
 import { TipBar } from '@/components/TipBar';
@@ -246,7 +247,10 @@ function UserDetailContent() {
         <PageContainer className="py-4">
           <div className="flex items-center justify-between gap-4">
             <MainNav days={days} />
-            <UserMenu />
+            <div className="flex items-center gap-3">
+              <UserMenu />
+              <MobileNav days={days} />
+            </div>
           </div>
         </PageContainer>
       </header>
